feat: implement dashboard components for agent health and applications overview, enhancing user experience on DAO dashboard page
