+++ apps/torus-governance/src/app/(pages)/dao-dashboard/_components/dashboard-tab/dashboard-redirect-card.tsx
